# swagger.yaml
swagger: '2.0'
info:
  description: DataCop REST API is to provide various information like alert, score and tests executed information for all datasets monitored by DataCop.
  version: 1.0.0
  title: DataCop REST API
  contact:
    email: datacopdev@microsoft.com
host: datacopprodfe.azurewebsites.net
basePath: /datacop/api/v1
...
schemes:
  - https
  - http
paths:
  /alerts:
    get:
      tags:
        - alerts
      summary: Get the alerts whose status are active or fired from now to cutofftime
      description: If not define cutofftime in query, the default cutofftime is seven days ago.
      operationId: getAllAlerts
      produces:
        - application/json
      parameters:
        - name: cutofftime
          in: query
          description: cutoff time stamp
          required: false
          type: string
          format: date-time
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/AlertDetailInfo'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - Bearer: []
  '/alerts/QoB/dataset/{datasetId}':
    get:
      tags:
        - alerts
      summary: Get the QoB alerts of a specific dataset
      description: Get all the QoB alerts for the specific dataset
      operationId: getAlertsBasedOnDatasetId
      produces:
        - application/json
      parameters:
        - name: datasetId
          in: path
          description: id of dataset
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/AlertDatasetInfo'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - Bearer: []
  '/alerts/QoB/dataset/kusto/{datasetId}':
    get:
      tags:
        - alerts
      summary: Get the QoB alerts of a specific dataset through Kusto
      description: Get all the QoB alerts for the specific dataset through Kusto
      operationId: getAlertsBasedOnDatasetIdByKusto
      produces:
        - application/json
      parameters:
        - name: datasetId
          in: path
          description: id of dataset
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/AlertDatasetInfoFromKusto'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - Bearer: []
...
securityDefinitions:
  Bearer:
    type: oauth2
    flow: application
    tokenUrl: 'https://login.microsoftonline.com/72f988bf-86f1-41af-91ab-2d7cd011db47/oauth2/token'
definitions:
...
  AlertDetailInfo:
    type: object
    properties:
      dateTime:
        type: string
        format: date-time
      severity:
        type: integer
        format: int32
        enum:
          - 0
          - 1
          - 2
          - 3
          - 4
      alertType:
        type: string
        enum:
          - None
          - Business
          - Data
      testType:
        type: string
        enum:
          - None
          - Availability
          - Latency
          - Correctness
          - Completeness
          - Custom
      datasetName:
        type: string
      incidentId:
        type: integer
        format: int64
      status:
        type: string
        enum:
          - Invalid
          - Active
          - Resolved
          - New
          - Mitigating
          - Correlating
          - Mitigated
          - Holding
          - Suppressed
      alertTitle:
        type: string
      titleOverride:
        type: string
      icmTeamName:
        type: string
      assignedTo:
        type: string
      acknowledgeDate:
        type: string
        format: date-time
      mitigateDate:
        type: string
        format: date-time
      resolvedDate:
        type: string
        format: date-time
      impactedDate:
        type: string
        format: date-time
  AlertDatasetInfo:
    type: object
    properties:
      id:
        type: integer
        format: int64
      title:
        type: string
      shortTitle:
        type: string
      owner:
        type: string
      severity:
        type: integer
        format: int32
        enum:
          - 0
          - 1
          - 2
          - 3
          - 4
      businessOwner:
        type: string
      status:
        type: string
        enum:
          - Invalid
          - Active
          - Resolved
          - New
          - Mitigating
          - Correlating
          - Mitigated
          - Holding
          - Suppressed
      alertType:
        type: string
        enum:
          - None
          - Business
          - Data
      createDate:
        type: string
        format: date-time
      acknowledgeDate:
        type: string
        format: date-time
      mitigateDate:
        type: string
        format: date-time
      resolveDate:
        type: string
        format: date-time
      impactStartDate:
        type: string
        format: date-time
  AlertDatasetInfoFromKusto:
    type: object
    properties:
      incidentId:
        type: integer
        format: int64
      title:
        type: string
      owningContactAlias:
        type: string
      alertStatus:
        type: string
        enum:
          - Invalid
          - Active
          - Resolved
          - New
          - Mitigating
          - Correlating
          - Mitigated
          - Holding
          - Suppressed
      issuedOnDate:
        type: string
        format: date-time
      mitigateDate:
        type: string
        format: date-time
      resolvedDate:
        type: string
        format: date-time
      lastUpdateDate:
        type: string
        format: date-time
      lastUpdate:
        type: string
...